tidy-check

+++ src/common/arrow/arrow_type_extension.cpp
@@ -326,7 +326,7 @@ struct ArrowBool8 {
 		auto source_ptr = reinterpret_cast<bool *>(source);
 		auto result_ptr = reinterpret_cast<int8_t *>(FlatVector::GetData(result));
 		for (idx_t i = 0; i < count; i++) {
-			result_ptr[i] = source_ptr[i];
+			result_ptr[i] = static_cast<int8_t>(source_ptr[i]);
 		}
 	}
 };
